Feat: UserService 기능 추출 및 SocialType 하드코딩 수정
